+++ sudoku_solver/__init__.py
@@ -1,0 +1,13 @@
+from .board import Board
+from .animation import animate_steps, _format_grid, GREEN, RED, YELLOW
+from . import example_boards
+
+__all__ = [
+    "Board",
+    "animate_steps",
+    "_format_grid",
+    "GREEN",
+    "RED",
+    "YELLOW",
+    "example_boards",
+]
